+++ src/types.jl
@@ -1,66 +1,31 @@
-# https://github.com/JuliaGeo/CommonDataModel.jl/blob/main/src/types.jl
-
-abstract type AbstractModel end
-abstract type AbstractProject <: AbstractModel end
-abstract type AbstractInstrument <: AbstractModel end
-abstract type AbstractProduct <: AbstractModel end
-abstract type AbstractDataSet <: AbstractProduct end
-
 """
-    Project <: AbstractProject
+    Registry(name, datasets; defaults=(;), metadata=NoMetadata(), kw...)
 
-A representation of a project or mission containing instruments and datasets.
-
-# Fields
-- `name`: The name of the project
-- `metadata`: Additional metadata
-- `instruments`: Collection of instruments
-- `datasets`: Collection of datasets
+A relation of [`Dataset`](@ref)s: rows sharing a selector vocabulary. 
+A mission, an instrument, or any other grouping is a `Registry`.
 """
-mutable struct Project{I, D, MD} <: AbstractProject
-    name::String
-    instruments::I
-    datasets::D
-    metadata::MD
-end
-
-"keyword-based constructor"
-function Project(; name="", instruments=Dict(), datasets=Dict(), metadata=NoMetadata(), kwargs...)
-    Project(name, instruments, datasets, merge(metadata, kwargs))
-end
-
-"""
-    Instrument <: AbstractInstrument
-
-# Fields
-- `name`: The name of the instrument
-- `metadata`: Additional metadata
-- `datasets`: Collection of datasets
-"""
-struct Instrument{D, MD} <: AbstractInstrument
+struct Registry{D,MD,K}
     name::String
     datasets::D
     metadata::MD
+    defaults::K
 end
 
-"keyword-based constructor"
-function Instrument(; name="", datasets=Dict(), metadata=NoMetadata(), kwargs...)
-    Instrument(name, datasets, merge(metadata, kwargs))
+function Registry(; name="", datasets=[], metadata=NoMetadata(), defaults=(;), kwargs...)
+    Registry(name, datasets, merge(metadata, kwargs), defaults)
 end
+Registry(name, datasets; kw...) = Registry(; name, datasets, kw...)
 
-"Construct an `Instrument` from a dictionary."
-Instrument(d::AbstractDict) = Instrument(; (Symbol(k) => v for (k, v) in d)...)
+Base.getindex(reg::Registry; kw...) = select(reg; kw...)
 
-Base.insert!(p::Project, i, v::AbstractInstrument) = (p.instruments[i] = v; p)
-Base.insert!(p::Union{Project,Instrument}, i, v::AbstractDataSet) = (p.datasets[i] = v; p)
-Base.push!(p::Union{Project,Instrument}, v) = insert!(p, name(v), v)
-Base.get(var::AbstractModel, s, d=nothing) = get(meta(var), s, d)
-Base.get(f::Function, var::AbstractModel, s) = get(f, meta(var), s)
+# https://spase-group.org/data/model/spase-2.7.0/spase-2_7_0_xsd.html#http___www.spase-group.org_data_schema_Spase_Catalog
+# Listing of events or observational notes.
+abstract type AbstractCatalog end
+abstract type AbstractEvent end
 
-for f in (:getindex, :iterate, :size, :length, :keys, :values)
-    @eval Base.$f(var::AbstractProduct, args...) = $f(parent(var), args...)
+@kwdef struct Event{A,T,M} <: AbstractEvent
+    data::A
+    start::T
+    stop::T
+    metadata::M = NoMetadata()
 end
-
-# https://github.com/rafaqz/DimensionalData.jl/blob/main/src/array/show.jl
-
-Base.show(io::IO, p::T) where {T<:AbstractModel} = print(io, name(p))